Imported GEOS version of FMS.
This is an orphan branch and should not ever be merged into the main
line of development.  Experiments indicate that the results are almost
bit identical to the master branch, except some roundoffs that show up
in the last bits of some isolated 64-bit checkpoints.  GEOS uses only
32 bits from FV, which presumably truncates these diffs most of the
time, but maybe not always.