Updates to allow building on macOS. Make include PRIVATE
Per @tclune, building with gfortran on macOS has some issues. Following
his lead with pFUnit, I've created a new file, `mojave_fix.h` and then
added the header to the C files that needed it.

Note, it looks like it has to be the *FIRST* header in a file that needs
it. Ugly, but what can you do.